Updated the login and profile components
